Added the struct that is going to be each piece.
Each square of the board should be a struct. Each struct (named square) should just be a tracker of where the pieces move.

So, instead of moving structs across the board, we simply have their ghost move across the board and we update the square they are standing on to make sure the board stays updated.

I am too tired to continue. Sleep <6h is not good.

I will implement functions tomorrow to calculate legal moves for each piece and also implement a way to make moves (and for the bot too). Then we can play against a rock tomorrow (random move engine).

- Finish legal moves.
- Create function to make moves.
- Create bot logic that makes random moves.
- Create check-mate logic. -> Probably... If no legal moves + king in check, then lose.

This will allow us to have our first match. From there, we should look to streamline the code a bit. From there we can create some actual logic decision.