I seperated the main.rs file into three distinct files.
I also changed the logic of how information propagates through the chess board. Now, we do not convert between values to assume a chess piece. Rather, we pass a struct (Piece) for each square. The downside is is that an empty square will still be considered "Piece" but to differentiate between an actual piece and an empty squre we have the value `is_empty` which will be the main logical component when determining whether a a square has a piece on it or not.

- perhaps in the future I will rename the struct "Square" and move it to the chess_board.rs file.