Added is_king_in_check function as alternative to is_pinned logic.
I have come to the realization that it would be far simpler to "allow" a piece to make an illegal move and simple check whether that put the king in harms way. For example, if a pawn was pinned through a bishop, we allow the pawn to make the move internally, and if the king is then in check as a result of the pawn's movement, we know the pawn was pinned.

Why are we doing it this way? Simply because the logic for live-checking a pin is very complex.

For example, suppose we want to check whether a pawn is pinned by a bishop on the top right diagonal.
- If there is another piece between pawn and king -> No pin.
- If there is another piece between pawn and bishop -> No pin.

However, this does not mean the pawn is frozen. The pawn could still capture the bishop, if permissable.

My current soltion simply says:
- If pawn moves AND king is now in check -> Pin exists.
- Everything else -> No pin.

Additionally, the king will act as a "reverse piece," where the king will check only the squares around him that COULD put him in check.
This means, we are not checking every single enemy piece if it's eyeing our king. But simply, the king is trying to EYE the enemy pieces.
- If the king can "see" an enemy piece -> King is check.