Compressed the diagonal check lookup.
I also found a bug in the checking mechanism for the rook and queen. Currently, we treat up/down/right/left as a single for loop for checking. Which means, if we are done checking upwards and we break, we exit the entire for loop and disregard the other directions.